+++ src/PostgREST/Plan.hs
@@ -352,7 +352,7 @@ initReadRequest ctx@ResolverContext{qi=QualifiedIdentifier{..}} =
   foldr (treeEntry rootDepth) $ Node defReadPlan{from=qi ctx, relName=qiName, depth=rootDepth} []
   where
     rootDepth = 0
-    defReadPlan = ReadPlan [] (QualifiedIdentifier mempty mempty) Nothing [] [] allRange mempty Nothing [] Nothing mempty Nothing Nothing False False [] rootDepth
+    defReadPlan = ReadPlan [] (QualifiedIdentifier mempty mempty) Nothing [] [] allRange mempty Nothing [] Nothing mempty Nothing Nothing Nothing False [] rootDepth
     treeEntry :: Depth -> Tree SelectItem -> ReadPlanTree -> ReadPlanTree
     treeEntry depth (Node si fldForest) (Node q rForest) =
       let nxtDepth = succ depth in
@@ -365,7 +365,7 @@ initReadRequest ctx@ResolverContext{qi=QualifiedIdentifier{..}} =
         SpreadRelation{..} ->
           Node q $
             foldr (treeEntry nxtDepth)
-            (Node defReadPlan{from=QualifiedIdentifier qiSchema selRelation, relName=selRelation, relHint=selHint, relJoinType=selJoinType, depth=nxtDepth, relIsSpread=True} [])
+            (Node defReadPlan{from=QualifiedIdentifier qiSchema selRelation, relName=selRelation, relHint=selHint, relJoinType=selJoinType, depth=nxtDepth, relSpread=Just selSpreadType} [])
             fldForest:rForest
         SelectField{..} ->
           Node q{select=CoercibleSelectField (resolveOutputField ctx{qi=from q} selField) selAggregateFunction selAggregateCast selCast selAlias:select q} rForest
@@ -437,7 +437,7 @@ expandStars ctx rPlanTree = Right $ expandStarsForReadPlan False rPlanTree
     adjustContext context fromQI _ = context{qi=fromQI}
 
 expandStarsForTable :: ResolverContext -> Bool -> ReadPlan -> ReadPlan
-expandStarsForTable ctx@ResolverContext{representations, outputType} hasAgg rp@ReadPlan{select=selectFields, relIsSpread=isSpread}
+expandStarsForTable ctx@ResolverContext{representations, outputType} hasAgg rp@ReadPlan{select=selectFields, relSpread=spreadType}
   -- We expand if either of the below are true:
   -- * We have a '*' select AND there is an aggregate function in this ReadPlan's sub-tree.
   -- * We have a '*' select AND the target table has at least one data representation.
@@ -451,6 +451,7 @@ expandStarsForTable ctx@ResolverContext{representations, outputType} hasAgg rp@R
     shouldExpandOrTag aggFunc = isNothing aggFunc || (isSpread && aggFunc == Just Count)
     hasDataRepresentation = any hasOutputRep knownColumns
     knownColumns = knownColumnsInContext ctx
+    isSpread = isJust spreadType
 
     hasOutputRep :: Column -> Bool
     hasOutputRep col = HM.member (colNominalType col, outputType) representations
@@ -511,8 +512,8 @@ addRels schema action allRels parentNode (Node rPlan@ReadPlan{relName,relHint,re
     updateForest rq = addRels schema action allRels rq `traverse` forest
 
 spreadRelIsNestedInToMany :: ReadPlan -> Bool
-spreadRelIsNestedInToMany ReadPlan{relIsSpread, relToParent, relIsInToManySpread} =
-  relIsSpread && (relIsInToManySpread || Just False == (relIsToOne <$> relToParent))
+spreadRelIsNestedInToMany ReadPlan{relSpread, relToParent, relIsInToManySpread} =
+  isJust relSpread && (relIsInToManySpread || Just False == (relIsToOne <$> relToParent))
 
 getJoinConditions :: Maybe Alias -> Maybe Alias -> Relationship -> [JoinCondition]
 getJoinConditions _ _ ComputedRelationship{} = []
@@ -625,7 +626,7 @@ findRel schema allRels origin target hint =
 -- * Are selected inside a to-many spread relationship
 -- * Are selected inside a to-one spread relationship but are nested inside a to-many spread relationship at any level
 addArrayAggToManySpread :: ReadPlanTree -> Either ApiRequestError ReadPlanTree
-addArrayAggToManySpread (Node rp@ReadPlan{select} forest) =
+addArrayAggToManySpread (Node rp@ReadPlan{select, relSpread} forest) =
     let newForest = addArrayAggToManySpread `traverse` forest
         newSelects
           | shouldAddArrayAgg = fieldToArrayAgg <$> select
@@ -635,7 +636,8 @@ addArrayAggToManySpread (Node rp@ReadPlan{select} forest) =
       shouldAddArrayAgg = spreadRelIsNestedInToMany rp
       fieldToArrayAgg field
         | isJust $ csAggFunction field = field
-        | otherwise = field { csAggFunction = Just ArrayAgg, csAlias = newAlias (csAlias field) (cfName $ csField field) }
+        | otherwise = field { csAggFunction = newAggFunc, csAlias = newAlias (csAlias field) (cfName $ csField field) }
+      newAggFunc = Just $ ArrayAgg $ relSpread == Just SpreadDistinct
       newAlias alias fieldName = maybe (Just fieldName) pure alias
 
 addRelSelects :: ReadPlanTree -> Either ApiRequestError ReadPlanTree
@@ -647,9 +649,9 @@ addRelSelects node@(Node rp forest)
     in Right $ Node rp { relSelect = newRelSelects } newForest
 
 generateRelSelectField :: ReadPlanTree -> Maybe RelSelectField
-generateRelSelectField (Node rp@ReadPlan{relToParent=Just _, relAggAlias, relIsSpread = True} _) =
+generateRelSelectField (Node rp@ReadPlan{relToParent=Just _, relAggAlias, relSpread = Just _} _) =
   Just $ Spread { rsSpreadSel = generateSpreadSelectFields rp, rsAggAlias = relAggAlias }
-generateRelSelectField (Node ReadPlan{relToParent=Just rel, select, relName, relAlias, relAggAlias, relIsSpread = False, relIsInToManySpread} forest) =
+generateRelSelectField (Node ReadPlan{relToParent=Just rel, select, relName, relAlias, relAggAlias, relSpread = Nothing, relIsInToManySpread} forest) =
   Just $ JsonEmbed { rsEmbedMode, rsSelName, rsAggAlias = relAggAlias, rsEmptyEmbed }
   where
     rsSelName = fromMaybe relName relAlias
@@ -663,7 +665,7 @@ generateRelSelectField (Node ReadPlan{relToParent=Just rel, select, relName, rel
 generateRelSelectField _ = Nothing
 
 generateSpreadSelectFields :: ReadPlan -> [SpreadSelectField]
-generateSpreadSelectFields rp@ReadPlan{select, relSelect} =
+generateSpreadSelectFields rp@ReadPlan{select, relSelect, relSpread} =
   -- We combine the select and relSelect fields into a single list of SpreadSelectField.
   selectSpread ++ relSelectSpread
   where
@@ -676,7 +678,7 @@ generateSpreadSelectFields rp@ReadPlan{select, relSelect} =
     relSelectToSpread :: RelSelectField -> [SpreadSelectField]
     relSelectToSpread (JsonEmbed{rsSelName}) =
       -- The regular embeds that are nested inside spread to-many relationships are also aggregated in an array
-      let (aggFun, alias) = if spreadRelIsNestedInToMany rp then (Just ArrayAgg, Just rsSelName) else (Nothing, Nothing) in
+      let (aggFun, alias) = if spreadRelIsNestedInToMany rp then (Just $ ArrayAgg $ relSpread == Just SpreadDistinct, Just rsSelName) else (Nothing, Nothing) in
       [SpreadSelectField { ssSelName = rsSelName, ssSelAggFunction = aggFun, ssSelAggCast = Nothing, ssSelAlias = alias }]
     relSelectToSpread (Spread{rsSpreadSel}) =
       rsSpreadSel
@@ -708,10 +710,10 @@ hoistSpreadAggFunctions :: ReadPlanTree -> Either ApiRequestError ReadPlanTree
 hoistSpreadAggFunctions tree = Right $ fst $ applySpreadAggHoistingToNode tree
 
 applySpreadAggHoistingToNode :: ReadPlanTree -> (ReadPlanTree, [HoistedAgg])
-applySpreadAggHoistingToNode (Node rp@ReadPlan{relAggAlias, relToParent, relIsSpread} children) =
+applySpreadAggHoistingToNode (Node rp@ReadPlan{relAggAlias, relToParent, relSpread} children) =
   let (newChildren, childAggLists) = unzip $ map applySpreadAggHoistingToNode children
       allChildAggLists = concat childAggLists
-      (newSelects, aggList) = if depth rp == 0 || (isJust relToParent && not relIsSpread)
+      (newSelects, aggList) = if depth rp == 0 || (isJust relToParent && isNothing relSpread)
                                 then (select rp, [])
                                 else hoistFromSelectFields relAggAlias (select rp)
 
@@ -719,11 +721,11 @@ applySpreadAggHoistingToNode (Node rp@ReadPlan{relAggAlias, relToParent, relIsSp
       -- child relationships, then it must hoist those aggregates to its parent rel.
       -- So we update them with the current `relAggAlias`.
       hoistAgg ((_, fieldName), hoistFunc) = ((relAggAlias, fieldName), hoistFunc)
-      hoistedAggList = if relIsSpread
+      hoistedAggList = if isJust relSpread
                        then aggList ++ map hoistAgg allChildAggLists
                        else aggList
 
-      newRelSelects = if null children || relIsSpread
+      newRelSelects = if null children || isJust relSpread
                       then relSelect rp
                       else map (hoistIntoRelSelectFields allChildAggLists) $ relSelect rp
   in  (Node rp { select = newSelects, relSelect = newRelSelects } newChildren, hoistedAggList)
@@ -839,7 +841,7 @@ addRelatedOrders (Node rp@ReadPlan{order,from} forest) = do
 --         relName = "projects",
 --         relToParent = Nothing,
 --         relJoinConds = [],
---         relAlias = Nothing, relAggAlias = "clients_projects_1", relHint = Nothing, relJoinType = Nothing, relIsSpread = False, relIsInToManySpread = False, depth = 1,
+--         relAlias = Nothing, relAggAlias = "clients_projects_1", relHint = Nothing, relJoinType = Nothing, relSpread = Nothing, relIsInToManySpread = False, depth = 1,
 --         relSelect = []
 --       },
 --       subForest = []
@@ -865,7 +867,7 @@ addRelatedOrders (Node rp@ReadPlan{order,from} forest) = do
 --         )
 --       ],
 --       order = [], range_ = fullRange, relName = "clients", relToParent = Nothing, relJoinConds = [], relAlias = Nothing, relAggAlias = "", relHint = Nothing,
---       relJoinType = Nothing, relIsSpread = False, relIsInToManySpread = False, depth = 0,
+--       relJoinType = Nothing, relSpread = Nothing, relIsInToManySpread = False, depth = 0,
 --       relSelect = []
 --     },
 --     subForest = subForst