+++ test/spec/Feature/Query/SpreadQueriesSpec.hs
@@ -125,6 +125,50 @@ aggDisabledSpec =
         , matchHeaders = [matchContentTypeJson]
         }
 
+    context "using the .. operator" $ do
+      it "works as an alias of the ... operator on a to-one spread" $ do
+        get "/projects?select=id,..clients(client_name:name)" `shouldRespondWith`
+          [json|[
+            {"id":1,"client_name":"Microsoft"},
+            {"id":2,"client_name":"Microsoft"},
+            {"id":3,"client_name":"Apple"},
+            {"id":4,"client_name":"Apple"},
+            {"id":5,"client_name":null}
+          ]|]
+          { matchStatus  = 200
+          , matchHeaders = [matchContentTypeJson]
+          }
+        get "/country?select=name,..capital(capital:name)" `shouldRespondWith`
+          [json|[
+            {"name":"Afghanistan","capital":"Kabul"},
+            {"name":"Algeria","capital":"Algiers"}
+          ]|]
+          { matchStatus  = 200
+          , matchHeaders = [matchContentTypeJson]
+          }
+
+      it "fails when it's a to-many relationship and aggregates are disabled" $ do
+        get "/clients?select=*,..projects(*)" `shouldRespondWith`
+          [json|{
+            "hint":null,
+            "details":null,
+            "code":"PGRST123",
+            "message":"Use of aggregate functions is not allowed"
+          }|]
+          { matchStatus  = 400
+          , matchHeaders = [matchContentTypeJson]
+          }
+        get "/supervisors?select=*,..processes(*)" `shouldRespondWith`
+          [json|{
+            "hint":null,
+            "details":null,
+            "code":"PGRST123",
+            "message":"Use of aggregate functions is not allowed"
+          }|]
+          { matchStatus  = 400
+          , matchHeaders = [matchContentTypeJson]
+          }
+
 aggEnabledSpec :: SpecWith ((), Application)
 aggEnabledSpec =
   describe "spread embeds" $ do
@@ -261,6 +305,100 @@ aggEnabledSpec =
           { matchStatus  = 200
           , matchHeaders = [matchContentTypeJson]
           }
+
+      context "using the .. operator to get distinct no null values" $ do
+        it "should work on a non-nested spread relationship" $ do
+          get "/factories?select=factory:name,..processes(category_id)&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory A","category_id":[1,2]},
+              {"factory":"Factory B","category_id":[1]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should return an empty array when no elements are found" $
+          get "/factories?select=factory:name,..processes(processes:name)&processes=is.null" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory D","processes":[]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work when selecting all columns" $
+          get "/factories?select=factory:name,..processes(*)&id=lte.2&order=name" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory A","id":[1,2],"name":["Process A1","Process A2"],"factory_id":[1],"category_id":[1,2]},
+              {"factory":"Factory B","id":[3,4],"name":["Process B1","Process B2"],"factory_id":[2],"category_id":[1]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested one-to-one relationship" $
+          get "/factories?select=factory:name,...processes(..process_costs(process_costs:cost))&id=eq.3" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory C","process_costs":[40.00,70.00]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested many-to-one relationship" $
+          get "/factories?select=factory:name,...processes(..process_categories(categories:name))&id=eq(any).{2,3}" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory B","categories":["Batch"]},
+              {"factory":"Factory C","categories":["Mass"]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested one-to-many relationship" $
+          get "/factories?select=factory:name,...processes(..process_supervisor(supervisor_ids:supervisor_id))&id=eq.3" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory C","supervisor_ids":[3]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested many-to-many relationship" $ do
+          get "/factories?select=factory:name,...processes(..supervisors(supervisors:name))&id=eq.3" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory C","supervisors":["Peter"]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread one-to-one relationship, returning an array of objects with distinct no-null values" $ do
+          get "/factories?select=factory:name,..processes(process_costs(cost))&id=eq.3" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory C","process_costs":[{"cost": 40.00},{"cost": 70.00}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread many-to-one relationship, returning an array of objects with distinct no-null values" $
+          get "/factories?select=factory:name,..processes(process_categories(name))&id=eq(any).{2,3}" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory B","process_categories":[{"name": "Batch"}]},
+              {"factory":"Factory C","process_categories":[{"name": "Mass"}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread one-to-many relationship, returning an array of objects with distinct no-null values" $
+          get "/factories?select=factory:name,..processes(process_supervisor(supervisor_id))&id=eq.3" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory C","process_supervisor":[{"supervisor_id": 3}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread many-to-many relationship, returning an array of objects with distinct no-null values" $
+          get "/factories?select=factory:name,..processes(supervisors(name))&id=eq.3" `shouldRespondWith`
+            [json|[
+              {"factory":"Factory C","supervisors":[{"name": "Peter"}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
 
     context "many-to-many relationships as array aggregates" $ do
       it "should aggregate a single spread column" $ do
@@ -404,3 +542,112 @@ aggEnabledSpec =
           { matchStatus  = 200
           , matchHeaders = [matchContentTypeJson]
           }
+
+      context "using the .. operator to get distinct no null values" $ do
+        it "should work on a non-nested spread relationship" $
+          get "/operators?select=operator:name,..processes(category_id)&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","category_id":[1,2]},
+              {"operator":"Louis","category_id":[1,2]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should return an empty array when no elements are found" $
+          get "/operators?select=operator:name,..processes(processes:name)&processes=is.null" `shouldRespondWith`
+            [json|[
+              {"operator":"Liz","processes":[]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work when selecting all columns" $
+          get "/operators?select=operator:name,..processes(*)&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","id":[1,2,4],"name":["Process A1","Process A2","Process B2"],"factory_id":[1,2],"category_id":[1,2]},
+              {"operator":"Louis","id":[1,2],"name":["Process A1","Process A2"],"factory_id":[1],"category_id":[1,2]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested one-to-one relationship" $
+          get "/operators?select=operator:name,...processes(..process_costs(process_costs:cost))&id=eq(any).{3,5}" `shouldRespondWith`
+            [json|[
+              {"operator":"Alfred","process_costs":[70.00]},
+              {"operator":"Jeff","process_costs":[70.00,180.00,200.00]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested many-to-one relationship" $
+          get "/operators?select=operator:name,...processes(..process_categories(categories:name))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","categories":["Batch","Mass"]},
+              {"operator":"Louis","categories":["Batch","Mass"]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested one-to-many relationship" $
+          get "/operators?select=operator:name,...processes(..process_supervisor(supervisor_ids:supervisor_id))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","supervisor_ids":[1,2]},
+              {"operator":"Louis","supervisor_ids":[1,2]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested many-to-many relationship" $
+          get "/operators?select=operator:name,...processes(..supervisors(supervisors:name))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","supervisors":["John","Mary"]},
+              {"operator":"Louis","supervisors":["John","Mary"]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread one-to-one relationship, returning an array of objects with distinct no-null values" $ do
+          get "/operators?select=operator:name,..processes(process_costs(cost))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","process_costs":[{"cost": 70.00},{"cost": 150.00},{"cost": 200.00}]},
+              {"operator":"Louis","process_costs":[{"cost": 150.00},{"cost": 200.00}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread many-to-one relationship, returning an array of objects with distinct no-null values" $
+          get "/operators?select=operator:name,..processes(process_categories(name))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","process_categories":[{"name": "Batch"},{"name": "Mass"}]},
+              {"operator":"Louis","process_categories":[{"name": "Batch"},{"name": "Mass"}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-sporead one-to-many relationship, returning an array of objects with distinct no-null values" $
+          get "/operators?select=operator:name,..processes(process_supervisor(supervisor_id))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","process_supervisor":[{"supervisor_id": 1},{"supervisor_id": 2}]},
+              {"operator":"Louis","process_supervisor":[{"supervisor_id": 1},{"supervisor_id": 2}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on a nested non-spread many-to-many relationship, returning an array of objects with distinct no-null values" $
+          get "/operators?select=operator:name,..processes(supervisors(name))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","supervisors":[{"name": "John"},{"name": "Mary"}]},
+              {"operator":"Louis","supervisors":[{"name": "John"},{"name": "Mary"}]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }
+        it "should work on both the nested and top spread relationship, returning distinct no-null values for both" $ do
+          get "/operators?select=operator:name,..processes(process:name,..supervisors(supervisors:name))&id=lte.2" `shouldRespondWith`
+            [json|[
+              {"operator":"Anne","process":["Process A1","Process A2","Process B2"],"supervisors":["John","Mary"]},
+              {"operator":"Louis","process":["Process A1","Process A2"],"supervisors":["John","Mary"]}
+            ]|]
+            { matchStatus  = 200
+            , matchHeaders = [matchContentTypeJson]
+            }